churn modelling
Implemented Churn machine learning model for a bank using a two hidden layer artificial neural networks .
Obtained an accuracy of 86.25% over the 2000 examples of the test set data.
Used Stochastic gradient descent with  with batch_size 10 and no of epochs  = 100 to train the nn model for 8000 examples.Adam optimization technique was also applied.